// src/app/interface/registra.interface.ts
export interface RegistraDati {
	nome: string;
	cognome: string;
	sesso: string;
	dataNascita: Date;
	nazione: number;
	citta: number;
	provincia: string;
	indirizzo: string;
	civico: string;
	cittadinanza: string;
	codFiscale: string;
	username: string;
	email: string;
	psw: string;
	preferito: string;
	checkDati: boolean;
}